+++ CHANGELOG.md
@@ -6,7 +6,7 @@
 
 ### Added
 
-* Added option for user to input list of countries, see config option `countryList`
+* Added option for user to input list of countries to display, see config option `countryList`.
 
 ### Changed